+++ examples/large_models/deepspeed/opt/custom_handler.py
@@ -6,7 +6,8 @@ import transformers
 from transformers import AutoModelForCausalLM, AutoTokenizer
 
 from ts.context import Context
-from ts.torch_handler.distributed.base_deepspeed import BaseDeepSpeedHandler
+from ts.handler_utils.distributed.deepspeed import get_ds_engine
+from ts.torch_handler.distributed.base_deepspeed_handler import BaseDeepSpeedHandler
 
 logger = logging.getLogger(__name__)
 logger.info("Transformers version %s", transformers.__version__)
@@ -34,7 +35,11 @@ class TransformersSeqClassifierHandler(BaseDeepSpeedHandler, ABC):
         self.max_length = int(ctx.model_yaml_config["handler"]["max_length"])
         self.model = AutoModelForCausalLM.from_pretrained(model_dir, use_cache=False)
         self.tokenizer = AutoTokenizer.from_pretrained(model_dir, return_tensors="pt")
-        super().initialize(ctx)
+
+        ds_engine = get_ds_engine(self.model, ctx)
+        self.model = ds_engine.module
+        logger.info("Model %s loaded successfully", ctx.model_name)
+        self.initialized = True
 
     def preprocess(self, requests):
         """
