+++ test/postman/large_model_inference_data.json
@@ -6,5 +6,13 @@
         "synchronous":"true",
         "file":"../examples/large_models/Huggingface_pippy/sample_text.txt",
         "expected": "Hey, are you conscious? Can you talk to me?\nailabi lity"
+    },
+    {
+        "url":"https://torchserve.s3.amazonaws.com/mar_files/opt-ds.tar.gz",
+        "model_name":"opt",
+        "worker":1,
+        "synchronous":"true",
+        "file":"../examples/large_models/Huggingface_pippy/sample_text.txt",
+        "expected": "Hey, are you conscious? Can you talk to me?\nailabi lity"
     }
 ]
